+++ NEWS.md
@@ -102,6 +102,7 @@ New library features
   the uniquing checking ([#53474])
 * `RegexMatch` objects can now be used to construct `NamedTuple`s and `Dict`s ([#50988])
 * `Lockable` is now exported ([#54595])
+* New `ltrunc`, `rtrunc` and `ctrunc` functions for truncating strings to text width, accounting for char widths ([#55351])
 
 Standard library changes
 ------------------------